[SDSAET-38] Move JsError, SourceError, ScreenError models to models-api module
