+++ api/models-api/src/main/java/com/cognifide/aet/models/jserrors/JsErrorWrapper.java
@@ -1,49 +1,41 @@
-/**
+/*
  * AET
- * <p>
+ *
  * Copyright (C) 2013 Cognifide Limited
- * <p>
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
- * <p>
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.models;
+package com.cognifide.aet.models.jserrors;
 
-import com.cognifide.aet.job.common.comparators.source.diff.ResultDelta;
+import com.cognifide.aet.job.api.collector.JsErrorLog;
+import com.cognifide.aet.models.ErrorWrapper;
 import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
-import java.util.List;
-import java.util.Map;
+import java.util.Set;
 
-public class SourceErrorWrapper implements ErrorWrapper {
+public class JsErrorWrapper implements ErrorWrapper {
 
-  public static final String COMPARATOR_TYPE = "source";
-  public static final Type ARTIFACT_TYPE = new TypeToken<Map<String, List<ResultDelta>>>() {
+  public static final Type ARTIFACT_TYPE = new TypeToken<Set<JsErrorLog>>() {
   }.getType();
 
-  private final Map<String, List<ResultDelta>> result;
+  private final Set<JsErrorLog> jsErrors;
   private final String urlName;
-  private final Map<String, String> data;
 
-  public SourceErrorWrapper(Map<String, List<ResultDelta>> result, String urlName,
-      Map<String, String> data) {
-    this.result = result;
+  public JsErrorWrapper(Set<JsErrorLog> jsErrors, String urlName) {
+    this.jsErrors = jsErrors;
     this.urlName = urlName;
-    this.data = data;
   }
 
-  public Map<String, List<ResultDelta>> getResult() {
-    return result;
-  }
-
-  public Map<String, String> getData() {
-    return data;
+  public Set<JsErrorLog> getJsErrors() {
+    return jsErrors;
   }
 
   @Override
