+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/source/SourceComparator.java
@@ -22,8 +22,8 @@ import com.cognifide.aet.job.api.comparator.ComparatorProperties;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
 import com.cognifide.aet.job.common.comparators.source.diff.DiffParser;
-import com.cognifide.aet.job.common.comparators.source.diff.ResultDelta;
-import com.cognifide.aet.job.common.comparators.source.diff.ResultDelta.TYPE;
+import com.cognifide.aet.models.source.ResultDelta;
+import com.cognifide.aet.models.source.ResultDelta.TYPE;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import java.io.IOException;
 import java.util.ArrayList;
