+++ rest-endpoint/src/main/java/com/cognifide/aet/repositories/ErrorWrapperRepository.java
@@ -17,14 +17,10 @@ package com.cognifide.aet.repositories;
 
 import com.cognifide.aet.communication.api.metadata.Comparator;
 import com.cognifide.aet.job.api.collector.JsErrorLog;
-import com.cognifide.aet.job.common.comparators.source.diff.ResultDelta;
 import com.cognifide.aet.job.common.comparators.statuscodes.StatusCodesComparatorResult;
 import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5ComparatorResult;
 import com.cognifide.aet.models.ErrorType;
 import com.cognifide.aet.models.ErrorWrapper;
-import com.cognifide.aet.models.JsErrorWrapper;
-import com.cognifide.aet.models.ScreenErrorWrapper;
-import com.cognifide.aet.models.SourceErrorWrapper;
 import com.cognifide.aet.models.StatusCodesErrorWrapper;
 import com.cognifide.aet.models.W3cHtml5ErrorWrapper;
 import com.cognifide.aet.models.accessibility.AccessibilityErrorWrapper;
@@ -33,6 +29,10 @@ import com.cognifide.aet.models.accessibility.AccessibilityReportConfiguration;
 import com.cognifide.aet.models.cookie.CookieCompareComparatorResult;
 import com.cognifide.aet.models.cookie.CookieErrorWrapper;
 import com.cognifide.aet.models.cookie.CookieTestComparatorResult;
+import com.cognifide.aet.models.jserrors.JsErrorWrapper;
+import com.cognifide.aet.models.screen.ScreenErrorWrapper;
+import com.cognifide.aet.models.source.ResultDelta;
+import com.cognifide.aet.models.source.SourceErrorWrapper;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.cognifide.aet.vs.DBKey;
 import java.io.IOException;
